Делает функции открытия и закрытия попапов
