+++ index.html
@@ -11,7 +11,7 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кексобукинг</title>
 </head>
-<body>
+<body class="page-body">
 
   <main>
     <div class="promo">
